mm: Increase Dirty Ratios These values are more optimal for N5!
Keep in mind: These values control how often the kernel writes data to disk. When your
apps write data to disk, Linux actually doesn't write the data out to the disk right
away, it actually writes the stuff to system memory and the kernel handles when and
how the data is actually going to be flushed to the disk. These values represent a
percentage, the higher the percentage, the longer it waits to flush, the lower the
percentage, the more often flushes will occur. Now remember, we are dealing with solid
state storage, not the traditional disk platter and spindle. So we are actually able
to delay flushes a little longer with solid state versus a traditional hard drive disk
which should increase battery life a little bit.

Signed-off-by: Chester Kener <Cl3Kener@gmail.com>
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>
Used commit: https://github.com/exynos7420/android_kernel_samsung_exynos7420/commit/500e61c0e41a0e97594766a9ac3a8cdb0776d798